Lab6/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
black==23.11.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.8.0
uuid==1.30
uvicorn==0.24.0.post1